speedometer-app/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2-binary==2.9.6
paho-mqtt==1.6.1
python-multipart==0.0.6
python-dotenv==1.0.0